Clean up some types
