+++ src/common/dom/apply_themes_on_element.ts
@@ -1,5 +1,6 @@
+import { Theme } from "../../data/ws-themes";
 import { darkStyles, derivedStyles } from "../../resources/styles";
-import { HomeAssistant, Theme } from "../../types";
+import { HomeAssistant } from "../../types";
 import {
   hex2rgb,
   lab2hex,
