+++ src/data/weather.ts
@@ -5,12 +5,39 @@ import {
   mdiWeatherRainy,
   mdiWeatherWindy,
 } from "@mdi/js";
+import {
+  HassEntityAttributeBase,
+  HassEntityBase,
+} from "home-assistant-js-websocket";
 import { css, html, svg, SVGTemplateResult, TemplateResult } from "lit-element";
 import { styleMap } from "lit-html/directives/style-map";
 import { formatNumber } from "../common/string/format_number";
 import "../components/ha-icon";
 import "../components/ha-svg-icon";
-import type { HomeAssistant, WeatherEntity } from "../types";
+import type { HomeAssistant } from "../types";
+
+interface ForecastAttribute {
+  temperature: number;
+  datetime: string;
+  templow?: number;
+  precipitation?: number;
+  precipitation_probability?: number;
+  humidity?: number;
+  condition?: string;
+  daytime?: boolean;
+}
+
+interface WeatherEntityAttributes extends HassEntityAttributeBase {
+  temperature: number;
+  humidity?: number;
+  forecast?: ForecastAttribute[];
+  wind_speed: string;
+  wind_bearing: string;
+}
+
+export interface WeatherEntity extends HassEntityBase {
+  attributes: WeatherEntityAttributes;
+}
 
 export const weatherSVGs = new Set<string>([
   "clear-night",
