+++ src/data/webhook.ts
@@ -5,6 +5,10 @@ export interface Webhook {
   domain: string;
   name: string;
 }
+export interface WebhookError {
+  code: number;
+  message: string;
+}
 
 export const fetchWebhooks = (hass: HomeAssistant): Promise<Webhook[]> =>
   hass.callWS({
