+++ src/dialogs/more-info/controls/more-info-group.ts
@@ -10,8 +10,9 @@ import {
 import { html, TemplateResult } from "lit-html";
 import { dynamicElement } from "../../../common/dom/dynamic-element-directive";
 import { computeStateDomain } from "../../../common/entity/compute_state_domain";
+import { GroupEntity } from "../../../data/group";
 import "../../../state-summary/state-card-content";
-import { GroupEntity, HomeAssistant } from "../../../types";
+import { HomeAssistant } from "../../../types";
 import {
   importMoreInfoControl,
   domainMoreInfoType,
