+++ src/panels/lovelace/components/hui-image.ts
@@ -15,9 +15,9 @@ import { styleMap } from "lit-html/directives/style-map";
 import { STATES_OFF } from "../../../common/const";
 import parseAspectRatio from "../../../common/util/parse-aspect-ratio";
 import "../../../components/ha-camera-stream";
-import { fetchThumbnailUrlWithCache } from "../../../data/camera";
+import { CameraEntity, fetchThumbnailUrlWithCache } from "../../../data/camera";
 import { UNAVAILABLE } from "../../../data/entity";
-import { CameraEntity, HomeAssistant } from "../../../types";
+import { HomeAssistant } from "../../../types";
 
 const UPDATE_INTERVAL = 10000;
 const DEFAULT_FILTER = "grayscale(100%)";
