+++ src/panels/lovelace/entity-rows/hui-media-player-entity-row.ts
@@ -28,6 +28,7 @@ import {
   SUPPORT_VOLUME_MUTE,
   SUPPORT_VOLUME_SET,
   computeMediaDescription,
+  MediaPlayerEntity,
 } from "../../../data/media-player";
 import type { HomeAssistant } from "../../../types";
 import { hasConfigOrEntityChanged } from "../common/has-changed";
@@ -80,7 +81,7 @@ class HuiMediaPlayerEntityRow extends LitElement implements LovelaceRow {
       return html``;
     }
 
-    const stateObj = this.hass.states[this._config.entity];
+    const stateObj = this.hass.states[this._config.entity] as MediaPlayerEntity;
     const state = stateObj.state;
 
     if (!stateObj) {
